Added a remove and reset button to generate data
In Submission_page.html, I've added a "Reset button" that comes with a
warning (called using showWarning() and hideWarning() in
XMLgenerator-v2.js) before executing the new resetForm() in
XMLgenerator-v2.js. I have also added a "Delete last entry" button that
removes the last entry in a form by calling the DeleteSection() in
XMLgenerator-v2.js.

Some UI changes were adding a padding to the button of the "Generate XML
download" that appears when all forms have been filled in. This was
added due to the buttons not having a shadow behind it, making it look
different from all buttons. I've decreased size of Excel conversion
modal so it looks more natural when opened in index.html as well as
modified the Excel conversion text box width so it fits the entire modal
using the changeWidthConvereter() found in Submission_page.html.

Style.css modifications include corresponding styling for the new
warning section when resetting the form.

In XMLgenerator-v2.js, I've modified the CloneSection() function by
making new cloned entries have empty input fields instead of being a
copy of the first field by calling the new resetLastEntryValues()
function.