Finished Excel template and fixed bugs
Finished creating the template for the Excel file as well as
protected/locked certain sheets.

Fixed a bug in generating new data from custom.js where, for an unknown
reason, the populate_table(status) was not calling webservice.cgi to
load new data. This was temporarily "fixed" using a setTimeout to force
call populate_table(status) again after 2.5 seconds

In Submission_page.html, removed "margin-top:1px" from the Reset form
styling. That was originally there for when the button had a different
red design. Forgot to remove it when decided to keep it consistent with
other buttons.

For XMLgenerator-v2.js, I changed the "respository type*" call name in
the convert_to_json() function to match the new Excel template. Also
fixed bug where Leaf's petiole was not appearing in the generated XMLs.