Fixed generating new data bugs
Fixed a bug that prevents uploaded data from being loaded as well as
made loading new data 26% faster.

The bug (temporarily fixed using setTimeout(function(){
populate_table(0); }, 2500); in custom.js) was caused by a failure in
the get_gene_structures.cgi taking too long to load. Therefore, I added
an error case that calls populate_table(status) and
populate_efp_modal(status) in the variants_radio_options(status)
function found in custom.js . With this, I've also added a warning which
replaces the splicing variants. This warning was coded by the
"warning_core" class in core\style.css (colours in
Submission_page\style.css's warning class was changed to match this
warning)

With this in mind, I removed the if call in custom.js's
populate_table(status) function that requires rnaseq_calls.length and
count_bam_entries_in_xml to be equal due to count_bam_num() is sometimes
slow to execute. With this removal, rnaseq_images(status) will always be
called whenever populate_table(status) is called.

Overall, the loading of a new data is now faster (as well as memory
required was reduced) by preventing xhr.send(null) to be called
continuously in count_bam_num() while the progress_percent is less than
100. Now it only calls up to 1.5 times count_bam_entries_in_xml (this
was because calling it once caused an error and I was not sure why so I
set it to this temporarily). Doing this causes the default dataset load
time go from 4.6min to 3.4 min (26.08% faster).

Work in incorporating Google Identity to submit data to the public
RNA-Seq Database has begun. Google Identity API script was added
including its meta-data and Content ID (which is registered by my Google
account: alexander.joohyun.sullivan@gmail.com). New button was created
which allows users to login in index.html. Using the new
onSignIn(googleUser) function, the image of the icon now changes to the
Google's user profile picture when signed in. Their email is now the
variable: users_email. This is going to be used to add in their RNA-Seq
databases to the list of options to load.

With new databases being added, xml_src_dic.js was created to be a list
of available databases in a JavaScript dictionary format. Changes to
index.html to match this was made including new way to call base_src in
change_dataset() function and uploaded database being pushed to this
dictionary during browser session.