Begin work on TMM normalization scripts
Begun work on creating scripts for client-side rendering of TMM
normalization of RNA-Seq data to be available for the users in addition
to RPKM/FPKM.

The entire TMM normalization math can be found in the
tmm_normalization.js file

In custom.js, changed error function so now when get_gene_structure.cgi
encounters an error, the entire table empties. Also fixed wording of
error message and displayed more buttons during error screen.

Fixed formatting in bulk_template.xlsx

In index.html, added script for tmm_normalization.js as well as math.js
which is used in TMM normalization as well as added new ID tag to
options and variants to correspond with custom.js updates.