Fixed bugs in generating user submitted data
Fixed bugs in generating user submitted data.

Based on the Excel template, if a user chooses a subunit of a tissue
that does not match its designated tissue, it will be given a broad
version of that SVG tissue’s subunit (most commonly referred to as
“all”). This can be see in custom.js’s new function checkSubunit(svg,
subunit) and called in populate_table(status). Noticed that in the Excel
template, people are mixing up leaf and Full_leaf; plan on removing
leaf.svg soon to reduce confusion.

In custom.js, work has begun on fixing a known bug where SRA during ajax
in rnaseq_image(status) becomes undefined. Currently trying to figure
out why this is happening and how to prevent it or get around it. An
error image was created (error.png) to replaced the images of the
RNA-Seq data that corresponds with this issue.

Regarding uploading user data, in Submission_page.html, Other.svg,
index.html and XMLgenerator-v2.js, the “Other” SVG is now functional as
well as its size has been corrected. The size for Stage 1-4 Leaf has
also been corrected.

In XMLgenerator-v2.js, fixed bug that caused the publication link and
SRA link to be directed to same webpage.

Minor work was done on the TMM normalization. R scripts were compared to
the client-side JS version developed for the eFP-Seq Browser and I
noticed that the function for calculating the library size was missing.
This was added in under the function calcLibsize(data, libsize) in
tmm_normalization.js

Regarding volunteers, helped Samaher with Excel template and Jenny with
the parser.

Known bugs:
• New data causing an error in webservice retrieval of SRA accession
numbers (mentioned above, currently working on fix)
• After generating new data, scrolling up doesn’t work unless you scroll
down again first
• Tooltip hover area for eFP Overview button is too wide, colliding with
Generate Data
• Help page still not completed

TODO:
• Fixed bugs
• Incorporate TMM
• Public database submission system