Fixed ovaries SVG in main page
Fixed ovaries SVG in main page

The ovaries subunit has been corrected to allow it to be coloured
instead of having the whole SVG coloured when the user selects the
ovaries subunit. These corrections are done using:
• custom.js with changes of checkSubunit(svg, subunit) from “ovaries” to
“Ovary_tissue”
• bamdata_Developmental_transcriptome.xml by correcting svg_subunits
• Submission_page.html by changing “ovaries” to “Ovary_tissue”
• index.html by changing the SVG code so it is corrected now

Increased visibility of the Stigma and Ovaries SVG by adding the st0
class, making stroke width darker and the overall image larger.

Other changes to custom.js was reverting the new rnaseq_image_url that
was used to bug-fix the webservice_gdrive.cgi error back to its old
code.

Removed temp.html as it was no longer needed as the live version of the
eFP-Seq Browser use an updated version of index.html

Known bugs:
• UI bugs out when resolution is less than 598pixels wide
• After generating new data, scrolling up doesn’t work unless users
scroll down first
• Does not work on IE, unknown reasons
• Help page still not completed
• Webservice_gdrive.cgi is not working for more than 5 datasets at once
due to error 403 by Google (error 1 in webservice)