Store and retrieve XMLs by Google account
Store and retrieve XMLs by Google account

The user can now store and retrieve XMLs they upload onto their Google
accounts. Currently all uploaded XMLs are uploaded onto a user Google’s
account if they are logged in. Later-on, the user will get to choose
which ones they want to upload to account and delete XMLs.

This was done by changes to custom.js and index.html and deletion of
xml_src_dic.js

The list of public database was removed (and deleted) from the
xml_src_dic.js and put into custom.js as the variable
dataset_dictionary.

The following function get_xml_list_output() retrieves a list of all
XMLs on the account and creates variables for the rest of custom.js and
index.html to use and display on the “RNA-Seq Databases” under “Private
database options”.

The variable base_dataset_dictionary and function
reset_database_options() in custom.js is used to reset all default
values when a user logs out or when there is a change in the user’s XMLs
server-side.

The function add_user_xml_by_upload() in custom.js is used to add a user
into the database or an uploaded file to an existing account.

The main page (index.html) was also adjusted to these new changes.

Also a small change in custom.js was to stop calling
populate_efp_modal() in unnecessary places.

Known bugs:
• UI bugs out when resolution is less than 598pixels wide
• After generating new data, scrolling up doesn’t work unless users
scroll down first
• Does not work on IE, unknown reasons
• Help page still not completed
• Webservice_gdrive.cgi is not working for more than 5 datasets at once
due to error 403 by Google (error 1 in webservice)