Fixed bug in eFP Overview not showing filtered data
Fixed bug in eFP Overview not showing appropriate filtered data

Fixed a bug in the eFP Overview that, when a data is filtered, the data
will not be displayed correctly according to the filter.

This was done by adding new conditions to the
populate_efp_moldal(status) function with the filtered_2d_rpkmNames list
and changing remainder_efp and ef_length variables from using
efp_rep_2d.length to tr_of_table.length.

All other changes including modifying the variable table_dl_str,
append_str for the SVG file, downloadIndexTable_base and all the new
filtered_2d lists as well as the for loops within
populate_efp_molda(status) and changing the name of download_tableCSV()
to download_XMLtableCSV() in custom.js as well as introducing a
hiddenDownloadModal and download_button button on index.html was in
preparations for an upcoming feature to download the download as CSV.

Also, spelling corrections were made on index.html.

Known bugs:
• UI bugs out when resolution is less than 598pixels wide
• After generating new data, scrolling up doesn’t work unless users
scroll down first
• Does not work on IE, unknown reasons
• Help page still not completed