Can now download public datasets as XML and CSV
Can now download public datasets as XML and CSV

Users can now download the publicly available datasets as XML or as a
CSV from the “Manage Data” select under “User Management”.

This was done by creating a new publicDatabaseDownload section in the
logoutModal (the Manage Data modal) where it displays all publicly
available datasets.

Due to this, created a new function in custom.js called
disableDeletePublic() which disables the “Delete XML” button when a
public database is checked off. Regardless, the user wouldn’t be able to
delete a public database without access anyways but removes the illusion
the user has the choice.

With this new feature, change all function counts related to the XML/CSV
download from “Manage Data” from title_list.length to a new int:
total_amount_of_datasets, which is a combination of the amount of
private and public datasets user-accessible.

Also, the dataset dictionaries and the datasets itself. In the
dictionaries, removed the “Developmental transcriptome – Plant J –
Sample” as we are no longer using it despite we still have the file for
it. In the datasets bamdata_amazon_links.xml and
bamdata_Developmental_transcriptome.xml, corrected the title for these
XMLs.

Known bugs:
• UI bugs out when resolution is less than 598pixels wide
• After generating new data, scrolling up doesn’t work unless users
scroll down first
• Does not work on IE, unknown reasons
• Help page still not completed
• Controls are not in either of the Excel exporters
• When selecting more than two datasets to add from existing data to
generate new data, it does not add to the generate new data modal