Precached loading of public databases
Precached loading of public databases

Created the scripts for loading public databases from precached data at any time for the two publicly available databases “Araport 11 RNA-seq data” and “Developmental transcriptome – Klepikova et al” while the locus is AT2G24270.

This was done using the functions changePublicData() and checkPreload() found in custom.js.

Known bug with this feature is that the three Google Drive files in the Klepikova data are not preloading, this needs to be investigated and fixed.

Known bugs:
• UI bugs out when resolution is less than 598pixels wide
• After generating new data, scrolling up doesn’t work unless users scroll down first
• Does not work on IE, unknown reasons
• Help page still not completed
• Controls are not in either of the Excel exporters
• When selecting more than two datasets to add from existing data to
generate new data, it does not add to the generate new data modal
• Google Drive files in Klepikova data are not precache properly