GFF tooltip added to splice variants
GFF tooltip added to splice variants

Unique IDs/GFFs have been added to the splice variants of each gene which identifies to which splice variant they are identified by (identified by a decimal number for example AT1G65480.2 opposed to AT1G65480.1).

The code for this can be found throughout custom. A new function was coded called getGFF(locusID) which calls and parses a PHP scripted created by Asher and Priyank during early stages of the eFP-Seq Browser development which contains the location of the unique identifier (“uniqueID”) which adds it to the list GFF_List. This is only called when the script is initially opened ($(document).ready(function()) and whenever the locus changes (as can be seen in the changes made to the function update_all_images(status) where a if statement was added to compare the current locus value opposed to the previous one).

The GFF tooltip was added within the variants_radio_option(status) dynamic rendering of the client’s page.

Small UI/UX changes were made as well including adding the SRR number to the information panel on the right, moving login next to custom data and increasing size of public database select.

Known bugs:
• UI bugs out when resolution is less than 598pixels wide
• After generating new data, scrolling up doesn’t work unless users scroll down first
• Does not work on IE, unknown reasons
• Help page still not completed
• Controls are not in either of the Excel exporters
• When selecting more than two datasets to add from existing data to
generate new data, it does not add to the generate new data modal
• Google Drive files in Klepikova data are not precache properly