Private datasets now load in proper order
Private datasets now load in proper order

Fixed a bug where the private datasets load in the wrong order, causing names to incorrectly match with the wrong datasets. This was done by moving the data_list.push from inside the get_user_XML_display() function to outside in its own create_data_list(size) function where size is equal the length of the amount of private datasets linked to an account.

Also, added new if conditions to xml_title names if to check if the private dataset names are the same as the public dataset names which can cause an issue with adding to the dataset_dictionary’s dictionary.

Known bugs:
• UI bugs out when resolution is less than 598pixels wide
• After generating new data, scrolling up doesn’t work unless users scroll down first
• Does not work on IE, unknown reasons
• Help page still not completed
• Controls are not in either of the Excel exporters
• When selecting more than two datasets to add from existing data to
generate new data, it does not add to the generate new data modal
• Google Drive files in Klepikova data are not precache properly