Google Sign-in and Private Database now works on new UI
Google Sign-in and Private Database now works on new UI

The Google sign-in and the private database options now work the new UI design. Changes can be seen through the new “Sign in” button on index.html which corresponds with functions in custom.js such as hiddenGoogleSignin() and remove_private_database() (which use to be on index.html) which allows users to sign-in and refresh data when sign out. Styling changes were made in style.css to correspond with the new button.

Bug fixes:
• Google sign-in and private databases now work on new UI

TODO:
• Fixed wording
• Correct precached data for Araport 11
• Make splicing variants part of the RNA table instead of above it
• EA2016 auto-completion library implementation
• Implement IGB viewer into the eFP-Seq Browser
• TMM or TPM additional normalization
• Expand creation of private databases to include adding from public databases and private databases into your creation
• “Tour UI”
• Back-to-top button

Known bugs:
• After generating new data, scrolling up doesn’t work unless users scroll down first
• Does not work on IE, unknown reasons
• Help page still not completed
• Controls are not in either of the Excel exporters
• When selecting more than two datasets to add from existing data to generate new data, it does not add to the generate new data modal
• GFF ID parse output not work on redesigned UI